+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+1.1.0 (201?-??-??)
+------------------
+
+  * added `bovigo\assert\predicate\each()`
+  * added `bovigo\assert\phpunit\PHPUnit_Framework_TestCase::assertContainsOnly()`
+  * added `bovigo\assert\phpunit\PHPUnit_Framework_TestCase::assertContainsOnlyInstancesOf()`
+  * added `bovigo\assert\phpunit\PHPUnit_Framework_TestCase::assertNotContainsOnly()`
+
+
 1.0.0 (2015-12-27)
 ------------------